+++ apps/devlog/src/util/common.ts
@@ -1,6 +1,6 @@
 export function shuffle<T>(array: T[]) {
   return array
-    .map((value) => ({ value, sortKey: Math.random() }))
+    .map((value) => ({ sortKey: Math.random(), value }))
     .sort((a, b) => a.sortKey - b.sortKey)
     .map(({ value }) => value)
 }