+++ packages/astro-fuse/tsup.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  dts: true,
+  entry: {
+    client: 'src/loadFuse.ts',
+    index: 'src/index.ts',
+  },
+  format: ['esm'],
+})