+++ packages/react-use-fusejs/prettier.config.mjs
@@ -1,0 +1,12 @@
+/**
+ * @see https://prettier.io/docs/en/configuration.html
+ * @type {import("prettier").Config}
+ */
+const config = {
+  semi: false,
+  singleQuote: true,
+  tabWidth: 2,
+  trailingComma: 'es5',
+}
+
+export default config